Fix WiFi scan, SSID parse-out and command in WPA crack
Commit fixes WiFi scan in WPA crack by utilizing functions we've already got.
This means that WiFi scan and parsing have been moved into wless_commands.
Also, SSID parse-out is fixed. I have thought of case when SSID has ',' in the
name and it wouldn't work.
Last, but not least, airodump-ng command is, hopefully, fixed. SSID with
white-space in the name was breaking it as probably would non-ASCII SSIDs.